add Qwen3VL SFT and GRPO support

+++ torchtune/modules/kv_cache.py
@@ -63,12 +63,12 @@ class KVCache(nn.Module):
 
         Example:
             >>> cache = KVCache(batch_size=2, num_kv_heads=4, max_seq_len=16, head_dim=32, dtype=torch.bfloat16)
-            >>> keys, values = torch.ones((2, 4, 8, 32)), torch.ones((2, 4, 8, 32))
+            >>> keys, values = torch.ones((2, 4, 8, 32)).to(torch.bfloat16), torch.ones((2, 4, 8, 32)).to(torch.bfloat16)
             >>> cache.update(keys, values)
             >>> # now positions 0 through 7 are filled
             >>> cache.size
             >>> 8
-            >>> keys, values = torch.ones((2, 4, 1, 32)), torch.ones((2, 4, 1, 32))
+            >>> keys, values = torch.ones((2, 4, 1, 32)).to(torch.bfloat16), torch.ones((2, 4, 1, 32)).to(torch.bfloat16)
             >>> cache.update(keys, values)
             >>> # this will fill at position 8
             >>> cache.size
